refactor(mdw-sync): replace MdwPlugin interface with Plugin interface and update related services and plugins for improved consistency and maintainability
