+++ src/mdw-sync/plugins/plugin.interface.ts
@@ -2,17 +2,21 @@ import { Tx } from '../entities/tx.entity';
 
 export { Tx };
 
-export interface MdwPluginFilter {
+export interface PluginFilter {
   type?: 'contract_call' | 'spend';
   contractIds?: string[];
   functions?: string[];
   predicate?: (tx: Partial<Tx>) => boolean;
 }
 
-export interface MdwPlugin {
+export interface Plugin {
   name: string;
+  /**
+   * When the plugin version changes, it will cause a full re-sync of the plugin transactions.
+   */
+  version: number;
   startFromHeight(): number;
-  filters(): MdwPluginFilter[];
+  filters(): PluginFilter[];
   onTransactionsSaved(txs: Partial<Tx>[]): Promise<void>;
   onReorg?(rollBackToHeight: number): Promise<void>;
 }
